requirements.txt
asgi-redis
asgiref
bayoo-docx
beautifulsoup4
bs4
celery
deathbycaptcha
df2gspread
dj-nexmo
Django
django-crispy-forms
django_comments_xtd
django-ipware
dnspython
docx2pdf
docx2txt
Flask
folium
getmac
gspread
gunicorn
jupyter-client
jupyterlab
jupyterlab-server
lxml
MarkupSafe
mysql-connector
mysql-connector-python
mysqlclient
nexmo
notebook
numpy
oauth2client
openpyxl
pandas
pdf2text
pdfcrowd
pdfkit
pdfminer
pdfminer.six
Pillow
pylint
pymongo
pyOpenSSL
python-docx
python-dotenv
python-pptx
psycopg2-binary
redis
requests
scipy
selenium
Shapely
textract
urllib3
whitenoise
xlrd
XlsxWriter
xlwt
